requirements.txt
boto3
crispy-bootstrap5==0.6
Django>=4.2.16,<=4.3
django-allauth
djangorestframework
django-cors-headers
django-jazzmin>=3.0.1,<=3.1
django-bootstrap-datepicker-plus
django-materializecss-form
djangorestframework-simplejwt
environs[django]==9.5.0
pydantic
calendly>=1.1.1,<=1.2
django-crispy-forms
django-phonenumber-field[phonenumbers]
dj-places>=5.2.1,<=5.3
django-storages[boto3]>=1.14.4,<=1.15
django-tables2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gunicorn
Pillow
psycopg
pytest